Added missing cast to avoid warnings.
git-svn-id: https://svn.opendtect.org/od/trunk@29838 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/Algo/array1dinterpol.cc
@@ -197,7 +197,7 @@ void PolyArray1DInterpol::extrapolate( bool start )
 	const float arrval = arr_->get( nextvalidix );
 	if ( !mIsUdf(arrval) )
 	{
-	    posidxs[positridx] = nextvalidix;
+	    posidxs[positridx] = (float) nextvalidix;
 	    vals[positridx] = arrval;
 	    positridx++;
 	}
